Add guard and validation to opening BlockView

+++ src/client/src/components/header/handlers/OnView.ts
@@ -1,9 +1,20 @@
-import { ViewType } from "../../../models/project";
+import { ViewType, VIEW_TYPE } from "../../../models/project";
 import { changeFlowView } from "../../../redux/store/flow/actions";
 import { removeSecondaryNode } from "../../../redux/store/secondaryNode/actions";
+import { GetSelectedNode } from "../../../helpers";
 import { updateBlockElements } from "../../../modules/explorer/redux/actions";
+import { setValidation } from "../../../redux/store/validation/actions";
+import { TextResources } from "../../../assets/text";
 
 const OnView = (view: ViewType, dispatch: any) => {
+  const selectedNode = GetSelectedNode();
+
+  // BlockView can only be opened when a node is selected
+  if (view === VIEW_TYPE.BLOCKVIEW && !selectedNode) {
+    dispatch(setValidation(false, TextResources.Validation_BlockView));
+    return;
+  }
+
   dispatch(removeSecondaryNode());
   dispatch(updateBlockElements([]));
   dispatch(changeFlowView(view));
